+++ src/utils/src/Waiter.php
@@ -9,12 +9,12 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
-namespace Hyperf\Utils\Coroutine\Waiter;
+namespace Hyperf\Utils;
 
 use Closure;
 use Hyperf\Engine\Channel;
-use Hyperf\Utils\Coroutine;
-use Hyperf\Utils\Coroutine\Waiter\Exception\WaitTimeoutException;
+use Hyperf\Utils\Exception\ExceptionThrower;
+use Hyperf\Utils\Exception\WaitTimeoutException;
 use Throwable;
 
 class Waiter
